Function `plotPseudospectra` for plotting pseudospectra of a matrix
Numerical tool to understand visually on the complex plane the possible
pseudospectra of a matrix under different tolerances epsilon. The pseudospectrum
is the spectrum of an operator the numbers that are "almost" eigenvalues.

Fix plot pseudospectra name inconsistency

test for derivative test and linscale

Re-adding section in readme of numeric tools after rewriting history from squash fails